# Improves thrift encoding performance by special-casing ascii