In many cases, annotation values are ascii strings. This special-cases
for this, improving thrift encoding performance dramatically.

This finally gets encoding overhead of a client span under 1 microsecond
  (on "my laptop")

Before:
```
Benchmark                                         Mode  Cnt  Score   Error  Units
CodecBenchmarks.writeClientSpan_thrift_zipkin     avgt   15  1.029 ± 0.041  us/op
CodecBenchmarks.writeLocalSpan_thrift_zipkin      avgt   15  0.741 ± 0.108  us/op
CodecBenchmarks.writeRpcSpan_thrift_zipkin        avgt   15  2.956 ± 0.128  us/op
CodecBenchmarks.writeRpcV6Span_thrift_zipkin      avgt   15  3.490 ± 0.104  us/op
```

After:
```
Benchmark                                         Mode  Cnt  Score   Error  Units
CodecBenchmarks.writeClientSpan_thrift_zipkin     avgt   15  0.779 ± 0.024  us/op
CodecBenchmarks.writeLocalSpan_thrift_zipkin      avgt   15  0.497 ± 0.026  us/op
CodecBenchmarks.writeRpcSpan_thrift_zipkin        avgt   15  2.018 ± 0.093  us/op
CodecBenchmarks.writeRpcV6Span_thrift_zipkin      avgt   15  2.407 ± 0.048  us/op
```